# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BMC Auto-Capture Executor API v0.2",
    "version": "0.2.6"
  },
  "paths": {
...
    "/executor/v1/contracts/{contract_id}": {
      "get": {
        "summary": "Get Contract Detail",
        "description": "Return detailed contract for a specific API endpoint.\n\nRead-only: does not read Excel, connect to services, or modify state.",
        "operationId": "get_contract_detail_executor_v1_contracts__contract_id__get",
        "parameters": [
          {
            "name": "contract_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Contract Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/executor/v1/config/excel:path": {
      "post": {
        "summary": "Set Latest Excel Path",
        "description": "Set latest Excel by executor-local path (local debug only).\n\nUses ExcelConfigStore.activate_from_local_path() which copies the\nfile into executor_state/configs/by_hash/ and writes latest.json.\nHistorical .runtime/configs/latest.xlsx files are migrated by\nExcelConfigStore when present.",
        "operationId": "set_latest_excel_path_executor_v1_config_excel_path_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExcelPathRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExcelConfigAcceptResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/executor/v1/config/excel": {
      "post": {
        "summary": "Upload Excel",
        "description": "Upload Excel file from remote Windows client (multipart/form-data).\n\nUses ExcelConfigStore.activate_from_upload() which saves to\nexecutor_state/configs/by_hash/ and atomically writes latest.json.\nDoes NOT write to .runtime/configs/latest.xlsx.",
        "operationId": "upload_excel_executor_v1_config_excel_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_excel_executor_v1_config_excel_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExcelConfigAcceptResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/executor/v1/plans/{plan_id}:run": {
      "post": {
        "summary": "Start Plan Run",
        "description": "Start a plan run. Request body validated via PlanRunRequest schema.",
        "operationId": "start_plan_run_executor_v1_plans__plan_id__run_post",
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Plan Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanRunRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanRunAcceptResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/executor/v1/plans/{plan_id}": {
      "get": {
        "summary": "Get Plan",
        "description": "Get plan summary.\n\nHandles both:\n  - numeric local plan IDs: no query params needed\n  - external plan IDs: requires excelHash query param",
        "operationId": "get_plan_executor_v1_plans__plan_id__get",
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plan Id"
            }
          },
          {
            "name": "excelHash",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Optional Excel SHA-256 hash. When provided, the query is validated against the plan's bound Excel hash.",
              "default": "",
              "title": "Excelhash"
            },
            "description": "Optional Excel SHA-256 hash. When provided, the query is validated against the plan's bound Excel hash."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/executor/v1/plans/{plan_id}/items": {
      "get": {
        "summary": "Get Plan Items Route",
        "description": "Get plan per-item details.\n\nHandles both int and external (string) plan IDs.\nExternal plans require excelHash query param.",
        "operationId": "get_plan_items_route_executor_v1_plans__plan_id__items_get",
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plan Id"
            }
          },
          {
            "name": "excelHash",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Optional Excel SHA-256 hash. When provided, the query is validated against the plan's bound Excel hash.",
              "default": "",
              "title": "Excelhash"
            },
            "description": "Optional Excel SHA-256 hash. When provided, the query is validated against the plan's bound Excel hash."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanItemsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/executor/v1/plans/{plan_id}/callbacks:retry": {
      "post": {
        "summary": "Retry Plan Callbacks",
        "operationId": "retry_plan_callbacks_executor_v1_plans__plan_id__callbacks_retry_post",
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plan Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CallbackRetryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CallbackRetryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/executor/v1/plans": {
      "post": {
        "summary": "Start External Plan",
        "description": "Start external plan. Service only needs excelHash + callback URL.",
        "operationId": "start_external_plan_executor_v1_plans_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExternalPlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanRunAcceptResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ExcelConfigAcceptResponse": {
        "properties": {
          "accepted": {
            "type": "boolean",
            "title": "Accepted"
          },
          "deviceCount": {
            "type": "integer",
            "title": "Devicecount",
            "default": 0
          },
          "enabledDeviceCount": {
            "type": "integer",
            "title": "Enableddevicecount",
            "default": 0
          },
          "taskCount": {
            "type": "integer",
            "title": "Taskcount",
            "default": 0
          },
          "enabledTaskCount": {
            "type": "integer",
            "title": "Enabledtaskcount",
            "default": 0
          },
          "filename": {
            "type": "string",
            "title": "Filename",
            "default": ""
          },
          "excelHash": {
            "type": "string",
            "title": "Excelhash",
            "default": ""
          },
          "sha256": {
            "type": "string",
            "title": "Sha256",
            "default": ""
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "accepted"
        ],
        "title": "ExcelConfigAcceptResponse"
      },
      "ExcelPathRequest": {
        "properties": {
          "excelPath": {
            "type": "string",
            "title": "Excelpath",
            "description": "Executor-local .xlsx path to activate"
          }
        },
        "type": "object",
        "required": [
          "excelPath"
        ],
        "title": "ExcelPathRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    }
  }
}